+++ modules/backend/controllers/Index.php
@@ -1,5 +1,6 @@
 <?php namespace Backend\Controllers;
 
+use Redirect;
 use BackendMenu;
 use Backend\Classes\Controller;
 use Backend\Widgets\ReportContainer;
@@ -15,7 +16,10 @@ class Index extends Controller
 {
     use \Backend\Traits\InspectableContainer;
 
-    public $requiredPermissions = ['backend.access_dashboard'];
+    /**
+     * @see checkPermissionRedirect()
+     */
+    public $requiredPermissions = [];
 
     /**
      * Constructor.
@@ -26,42 +30,28 @@ class Index extends Controller
 
         BackendMenu::setContextOwner('October.Backend');
         new ReportContainer($this);
-
-        /* @todo Remove line if year >= 2015 */
-        if (\Schema::hasColumn('backend_users', 'activated')) {
-            \Schema::table('backend_users', function ($table) {
-                $table->renameColumn('activated', 'is_activated');
-            });
-        }
-        /* @todo Remove line if year >= 2015 */
-        if (\Schema::hasColumn('backend_user_throttle', 'suspended')) {
-            \Schema::table('backend_user_throttle', function ($table) {
-                $table->renameColumn('suspended', 'is_suspended');
-            });
-        }
-        /* @todo Remove line if year >= 2015 */
-        if (\Schema::hasColumn('backend_user_throttle', 'banned')) {
-            \Schema::table('backend_user_throttle', function ($table) {
-                $table->renameColumn('banned', 'is_banned');
-            });
-        }
-        /* @todo Remove line if year >= 2015 */
-        if (\Schema::hasColumn('deferred_bindings', 'bind')) {
-            \Schema::table('deferred_bindings', function ($table) {
-                $table->renameColumn('bind', 'is_bind');
-            });
-        }
-        /* @todo Remove line if year >= 2015 */
-        if (\Schema::hasColumn('system_files', 'public')) {
-            \Schema::table('system_files', function ($table) {
-                $table->renameColumn('public', 'is_public');
-            });
-        }
     }
 
     public function index()
     {
+        if ($redirect = $this->checkPermissionRedirect())
+            return $redirect;
+
         $this->pageTitle = 'backend::lang.dashboard.menu_label';
         BackendMenu::setContextMainMenu('dashboard');
+    }
+
+    /**
+     * Custom permissions check that will redirect to the next
+     * available menu item, if permission to this page is denied.
+     */
+    protected function checkPermissionRedirect()
+    {
+        if (!$this->user->hasAccess('backend.access_dashboard')) {
+            $true = function(){ return true; };
+            if ($first = array_first(BackendMenu::listMainMenuItems(), $true)) {
+                return Redirect::intended($first->url);
+            }
+        }
     }
 }