+++ modules/system/lang/en/lang.php
@@ -248,6 +248,7 @@ return [
         'manage_system_settings' => 'Manage system settings',
         'manage_software_updates' => 'Manage software updates',
         'manage_mail_templates' => 'Manage mail templates',
+        'manage_mail_settings' => 'Manage mail settings',
         'manage_other_administrators' => 'Manage other administrators',
         'view_the_dashboard' => 'View the dashboard'
     ]